Cleanup dash boards, all metrics now as expected
