+++ eth_consensus/compositions/composition.toml
@@ -12,7 +12,7 @@ name = "eth_consensus"
 
 [[groups]]
   id = "main-instances"
-  instances = { count = 3 }
+  instances = { count = 20 }
 
   [groups.run]
   # Check the plan (./eth_consensus/manifest.toml) to understand the meaning of each param
