+++ scripts/run_eth_consensus.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+
+# Run this from the root directory of the repository
+# i.e ./scripts/run_eth_consensus.sh
+
+cd eth_consensus
+cargo clean
+cd ..
+testground run composition -f ./eth_consensus/compositions/composition.toml --wait
